Removing manual addition of rules and automating the same. The engine can pick up rules added to rulesImpl/
Files changed :
1. dist file
2. Grunt to exclude glob, transform with req-glob and use index.js with
standalone set to capture global for window.
3. renamed a11y-auditor to index.js
4. Updated package.json with dev deps for req-glob and deps for glob.
Updated main file, keywords
5. Removed a11y-auditor.js main file
6. No need to manually add entries to rules. req-glob transform for
browserify to work when browserify builds are used - cheats the static
req analyser of browserify to expand dynamically to glob patterns and
when requiring the proj as a node module, the above setup will force
fail with an exception and will use glob module instead.