adding the elementVisibility checker and the event listener tracker script to the auditRunner
These get hooked on to the HTMLElement type and Node type and are
available across elements. They just have to be accessed as
elem.isVisible() and elem.getEventListener(eventType). To the
visibility checker script, although we determine if the element is off
the viewport, it could still be well within the scrollable area of the
page. This must be caclulated next.